# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrossBorder AI Demo API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/crowdsource/feed": {
      "get": {
        "tags": [
          "众包反馈"
        ],
        "summary": "获取近期持久化众包反馈",
        "operationId": "crowdsource_feed_api_crowdsource_feed_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrowdsourceFeedResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/crowdsource/report": {
      "post": {
        "tags": [
          "众包反馈"
        ],
        "summary": "提交并持久化众包反馈",
        "operationId": "submit_crowdsource_report_api_crowdsource_report_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrowdsourceReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrowdsourceSubmitResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/commercial/plans": {
      "get": {
        "tags": [
          "商业化演示"
        ],
        "summary": "获取本地商业套餐",
        "operationId": "get_plans_api_commercial_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialPlansResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/commercial/subscription": {
      "get": {
        "tags": [
          "商业化演示"
        ],
        "summary": "获取当前 Demo 账户商业订阅",
        "operationId": "get_subscription_api_commercial_subscription_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialSubscriptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/commercial/checkout": {
      "post": {
        "tags": [
          "商业化演示"
        ],
        "summary": "执行无真实扣款的模拟结账",
        "operationId": "checkout_api_commercial_checkout_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommercialCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialCheckoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/commercial/subscription/cancel": {
      "post": {
        "tags": [
          "商业化演示"
        ],
        "summary": "取消本地商业订阅",
        "operationId": "cancel_subscription_api_commercial_subscription_cancel_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialSubscriptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/subscription": {
      "post": {
        "tags": [
          "提醒订阅"
        ],
        "summary": "通过已弃用的单数路径创建订阅",
        "operationId": "create_subscription_legacy_api_subscription_post",
        "deprecated": true,
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRecord"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertPreferences": {
        "properties": {
          "advance_reminder": {
            "type": "boolean",
            "title": "Advance Reminder",
            "default": true
          },
          "anomaly_alert": {
            "type": "boolean",
            "title": "Anomaly Alert",
            "default": true
          },
          "better_route_alert": {
            "type": "boolean",
            "title": "Better Route Alert",
            "default": true
          }
        },
        "type": "object",
        "title": "AlertPreferences"
      },
...
      "CommercialCheckoutRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "enum": [
              "starter",
              "professional",
              "enterprise"
            ],
            "title": "Plan Id"
          },
          "billing_cycle": {
            "type": "string",
            "enum": [
              "monthly",
              "yearly"
            ],
            "title": "Billing Cycle"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "billing_cycle"
        ],
        "title": "CommercialCheckoutRequest"
      },
      "CommercialCheckoutResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "subscription": {
            "$ref": "#/components/schemas/CommercialSubscription"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "subscription",
          "message"
        ],
        "title": "CommercialCheckoutResponse"
      },
      "CommercialPlan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "monthly_price_hkd": {
            "type": "integer",
            "title": "Monthly Price Hkd"
          },
          "yearly_price_hkd": {
            "type": "integer",
            "title": "Yearly Price Hkd"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "highlighted": {
            "type": "boolean",
            "title": "Highlighted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "audience",
          "monthly_price_hkd",
          "yearly_price_hkd",
          "description",
          "features"
        ],
        "title": "CommercialPlan"
      },
      "CommercialPlansResponse": {
        "properties": {
          "plans": {
            "items": {
              "$ref": "#/components/schemas/CommercialPlan"
            },
            "type": "array",
            "title": "Plans"
          },
          "demo_notice": {
            "type": "string",
            "title": "Demo Notice"
          }
        },
        "type": "object",
        "required": [
          "plans",
          "demo_notice"
        ],
        "title": "CommercialPlansResponse"
      },
      "CommercialSubscription": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "persona_id": {
            "type": "string",
            "title": "Persona Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "billing_cycle": {
            "type": "string",
            "title": "Billing Cycle"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "price_hkd": {
            "type": "integer",
            "title": "Price Hkd"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "renews_at": {
            "type": "string",
            "format": "date-time",
            "title": "Renews At"
          },
          "receipt_id": {
            "type": "string",
            "title": "Receipt Id"
          },
          "demo_payment": {
            "type": "boolean",
            "title": "Demo Payment",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "persona_id",
          "organization_id",
          "plan_id",
          "plan_name",
          "billing_cycle",
          "status",
          "price_hkd",
          "started_at",
          "renews_at",
          "receipt_id"
        ],
        "title": "CommercialSubscription"
      },
      "CommercialSubscriptionResponse": {
        "properties": {
          "subscription": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommercialSubscription"
              },
              {
                "type": "null"
              }
            ]
          },
          "demo_notice": {
            "type": "string",
            "title": "Demo Notice"
          }
        },
        "type": "object",
        "required": [
          "demo_notice"
        ],
        "title": "CommercialSubscriptionResponse"
      },
      "CrossingChannel": {
        "type": "string",
        "enum": [
          "traveller",
          "vehicle",
          "cargo"
        ],
        "title": "CrossingChannel"
      },
      "CrowdLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "CrowdLevel"
      },
      "CrowdsourceFeedResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/CrowdsourceRecord"
            },
            "type": "array",
            "title": "Reports"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "reports",
          "total"
        ],
        "title": "CrowdsourceFeedResponse"
      },
      "CrowdsourceRecord": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "default": "demo-user"
          },
          "port": {
            "type": "string",
            "minLength": 1,
            "title": "Port"
          },
          "actual_wait_time": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Actual Wait Time"
          },
          "crowd_level": {
            "$ref": "#/components/schemas/CrowdLevel"
          },
          "comment": {
            "type": "string",
            "maxLength": 160,
            "title": "Comment",
            "default": ""
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Port Id"
          },
          "direction": {
            "$ref": "#/components/schemas/TravelDirection",
            "default": "hong_kong_to_shenzhen"
          },
          "channel": {
            "$ref": "#/components/schemas/CrossingChannel",
            "default": "traveller"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "time_label": {
            "type": "string",
            "title": "Time Label"
          },
          "quality_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Quality Score"
          },
          "quality_level": {
            "$ref": "#/components/schemas/ReportQualityLevel"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "used_for_prediction": {
            "type": "boolean",
            "title": "Used For Prediction"
          },
          "source_type": {
            "$ref": "#/components/schemas/ObservationSource"
          }
        },
        "type": "object",
        "required": [
          "port",
          "actual_wait_time",
          "crowd_level",
          "id",
          "timestamp",
          "time_label",
          "quality_score",
          "quality_level",
          "expires_at",
          "used_for_prediction",
          "source_type"
        ],
        "title": "CrowdsourceRecord"
      },
      "CrowdsourceReport": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "default": "demo-user"
          },
          "port": {
            "type": "string",
            "minLength": 1,
            "title": "Port"
          },
          "actual_wait_time": {
            "type": "integer",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Actual Wait Time"
          },
          "crowd_level": {
            "$ref": "#/components/schemas/CrowdLevel"
          },
          "comment": {
            "type": "string",
            "maxLength": 160,
            "title": "Comment",
            "default": ""
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Port Id"
          },
          "direction": {
            "$ref": "#/components/schemas/TravelDirection",
            "default": "hong_kong_to_shenzhen"
          },
          "channel": {
            "$ref": "#/components/schemas/CrossingChannel",
            "default": "traveller"
          }
        },
        "type": "object",
        "required": [
          "port",
          "actual_wait_time",
          "crowd_level"
        ],
        "title": "CrowdsourceReport"
      },
      "CrowdsourceSubmitResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "points_earned": {
            "type": "integer",
            "title": "Points Earned"
          },
          "model_updated": {
            "type": "boolean",
            "title": "Model Updated"
          },
          "report": {
            "$ref": "#/components/schemas/CrowdsourceRecord"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "forecast_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastFeedbackLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "calibration_preview": {
            "additionalProperties": true,
            "type": "object",
            "title": "Calibration Preview"
          }
        },
        "type": "object",
        "required": [
          "success",
          "points_earned",
          "model_updated",
          "report",
          "message",
          "calibration_preview"
        ],
        "title": "CrowdsourceSubmitResponse"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "user_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Action"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id",
          "category",
          "retryable"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "ForecastFeedbackLink": {
        "properties": {
          "forecast_run_id": {
            "type": "string",
            "title": "Forecast Run Id"
          },
          "forecast_port_id": {
            "type": "string",
            "title": "Forecast Port Id"
          },
          "linked": {
            "type": "boolean",
            "title": "Linked"
          },
          "calibration_linked": {
            "type": "boolean",
            "title": "Calibration Linked"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "forecast_run_id",
          "forecast_port_id",
          "linked",
          "calibration_linked"
        ],
        "title": "ForecastFeedbackLink"
      },
...
      "ReportQualityLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "ReportQualityLevel"
      },
...
      "SubscriptionRecord": {
        "properties": {
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "routine": {
            "$ref": "#/components/schemas/Routine"
          },
          "alerts": {
            "$ref": "#/components/schemas/AlertPreferences"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "next_alert": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Alert"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "subscription_id",
          "user_id",
          "routine",
          "alerts",
          "created_at",
          "updated_at"
        ],
        "title": "SubscriptionRecord"
      },
      "SubscriptionRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "default": "demo-user"
          },
          "routine": {
            "$ref": "#/components/schemas/Routine"
          },
          "alerts": {
            "$ref": "#/components/schemas/AlertPreferences"
          }
        },
        "type": "object",
        "required": [
          "routine"
        ],
        "title": "SubscriptionRequest"
      },
...
      "TravelDirection": {
        "type": "string",
        "enum": [
          "hong_kong_to_shenzhen",
          "shenzhen_to_hong_kong"
        ],
        "title": "TravelDirection"
      },
...
    }
  }
}